% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortune.R
\name{fortune}
\alias{fortune}
\title{Unix fortune}
\usage{
fortune(include_obscene = FALSE)
}
\arguments{
\item{include_obscene}{Whether to include obscene fortunes, TRUE or FALSE}
}
\value{
Invisible NULL
}
\description{
This function prints a random fortune from the fortunes package.
}
