% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fortune_data}
\alias{fortune_data}
\title{Data set for the 50 fortunes options.}
\format{
\describe{
\item{fortune_data}{The fortune strings.}
}
}
\source{
\url{http://www.fortunecookiemessage.com/archive.php?start=0}
}
\usage{
fortune_data
}
\description{
Data set for the 50 fortunes options.
}
\keyword{datasets}
