% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortunes.R
\name{fortunes}
\alias{fortunes}
\title{Fortune Teller}
\usage{
fortunes()
}
\value{
A string that contains your fortune
}
\description{
This function has no function argument. Call the function and follow the interactive instructions to get a fortune.
}
\examples{
fortunes()
}
