% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortunes.R
\name{fortunes}
\alias{fortunes}
\title{Fortune Teller}
\usage{
fortunes(fortuney_words = NULL, awesome_words = NULL, evil_words = NULL)
}
\arguments{
\item{fortuney_words}{description}

\item{awesome_words}{description}

\item{evil_words}{description}
}
\value{
A string that contains your fortune
}
\description{
This function has no function argument. Call the function and follow the interactive instructions to get a fortune.
}
\examples{
fortunes()
}
